Add getbyid to mocksetup for check for null in handler. Add delete in controller and implement in generic repository

+++ Learner.Application/Features/HandleExercises/Delete/DeleteExerciseRequestHandler.cs
@@ -1,13 +1,18 @@
 ﻿using Learner.Application.Contracts.Repos;
 using MediatR;
 
-namespace Learner.Application.Tests.ExercisesTests.DeleteExerciseTest;
+namespace Learner.Application.Features.HandleExercises.Delete;
 
 public class DeleteExerciseRequestHandler(IExerciseRepository exerciseRepository) : IRequestHandler<DeleteExerciseRequest, Unit>
 {
     public async Task<Unit> Handle(DeleteExerciseRequest request, CancellationToken cancellationToken)
     {
-        await exerciseRepository.DeleteAsync(request.Id);
+        var exercise = await exerciseRepository.GetByIdAsync(request.Id);
+
+        if (exercise != null)
+        {
+            await exerciseRepository.DeleteAsync(request.Id);
+        }
             
         return Unit.Value;
     }
